+++ lib/twilio/device.ts
@@ -18,9 +18,10 @@ import {
 } from './errors';
 import Log from './log';
 import {
-  defaultRegion,
+  getChunderURI,
   getRegionShortcode,
-  getRegionURI,
+  Region,
+  regionToEdge,
 } from './regions';
 import { Exception, queryToJson } from './util';
 
@@ -273,6 +274,11 @@ class Device extends EventEmitter {
   private _connectionSinkIds: string[] = ['default'];
 
   /**
+   * The name of the edge the {@link Device} is connected to.
+   */
+  private _edge: string | null = null;
+
+  /**
    * Whether each sound is enabled.
    */
   private _enabledSounds: Record<Device.ToggleableSound, boolean> = {
@@ -328,7 +334,6 @@ class Device extends EventEmitter {
     iceServers: [],
     noRegister: false,
     pStreamFactory: PStream,
-    region: defaultRegion,
     rtcConstraints: { },
     soundFactory: Sound,
     sounds: { },
@@ -492,6 +497,14 @@ class Device extends EventEmitter {
   }
 
   /**
+   * Returns the {@link Edge} value the {@link Device} is currently connected
+   * to. The value will be `null` when the {@link Device} is offline.
+   */
+  get edge(): string | null {
+    return this._edge;
+  }
+
+  /**
    * Set a handler for the error event.
    * @deprecated Use {@link Device.on}.
    * @param handler
@@ -532,6 +545,10 @@ class Device extends EventEmitter {
    * if not connected.
    */
   region(): string {
+    this._log.warn(
+      '`Device.region` is deprecated and will be removed in the next major ' +
+      'release. Please use `Device.edge` instead.',
+    );
     this._throwUnlessSetup('region');
     return typeof this._region === 'string' ? this._region : 'offline';
   }
@@ -580,6 +597,22 @@ class Device extends EventEmitter {
       throw new InvalidArgumentError('Token is required for Device.setup()');
     }
 
+    Object.assign(this.options, options);
+
+    this._log.setDefaultLevel(
+      this.options.debug
+        ? Log.levels.DEBUG
+        : this.options.warnings
+          ? Log.levels.WARN
+          : Log.levels.SILENT,
+    );
+
+    const regionURI = getChunderURI(
+      this.options.edge,
+      this.options.region,
+      this._log.warn.bind(this._log),
+    );
+
     if (typeof Device._isUnifiedPlanDefault === 'undefined') {
       Device._isUnifiedPlanDefault = typeof window !== 'undefined'
         && typeof RTCPeerConnection !== 'undefined'
@@ -613,15 +646,9 @@ class Device extends EventEmitter {
 
     this.isInitialized = true;
 
-    Object.assign(this.options, options);
-
     if (this.options.dscp) {
       (this.options.rtcConstraints as any).optional = [{ googDscp: true }];
     }
-
-    this._log.setDefaultLevel(this.options.debug ? Log.levels.DEBUG
-      : this.options.warnings ? Log.levels.WARN
-      : Log.levels.SILENT);
 
     const getOrSetSound = (key: Device.ToggleableSound, value?: boolean) => {
       if (!hasBeenWarnedSounds) {
@@ -641,10 +668,6 @@ class Device extends EventEmitter {
         .forEach((eventName: Device.SoundName) => {
       this.sounds[eventName] = getOrSetSound.bind(null, eventName);
     });
-
-    const regionURI = getRegionURI(this.options.region, newRegion => this._log.warn(
-      `Region ${this.options.region} is deprecated, please use ${newRegion}.`,
-    ));
 
     this.options.chunderw = `wss://${this.options.chunderw || regionURI}/signal`;
 
@@ -838,10 +861,9 @@ class Device extends EventEmitter {
       ice_restart_enabled: this.options.enableIceRestart,
       platform: rtc.getMediaEngine(),
       sdk_version: C.RELEASE_VERSION,
-      selected_region: this.options.region,
     };
 
-    function setIfDefined(propertyName: string, value: string | undefined) {
+    function setIfDefined(propertyName: string, value: string | undefined | null) {
       if (value) { payload[propertyName] = value; }
     }
 
@@ -853,11 +875,11 @@ class Device extends EventEmitter {
       payload.direction = connection.direction;
     }
 
-    const stream: IPStream = this.stream;
-    if (stream) {
-      setIfDefined('gateway', stream.gateway);
-      setIfDefined('region', stream.region);
-    }
+    setIfDefined('gateway', this.stream && this.stream.gateway);
+    setIfDefined('selected_region', this.options.region);
+    setIfDefined('region', this.stream && this.stream.region);
+    setIfDefined('selected_edge', this.options.edge);
+    setIfDefined('edge', this._edge || this._region);
 
     return payload;
   }
@@ -1018,7 +1040,9 @@ class Device extends EventEmitter {
    * Called when a 'connected' event is received from the signaling stream.
    */
   private _onSignalingConnected = (payload: Record<string, any>) => {
-    this._region = getRegionShortcode(payload.region) || payload.region;
+    const region = getRegionShortcode(payload.region);
+    this._edge = regionToEdge[region as Region] || payload.region;
+    this._region = region || payload.region;
     this._sendPresence();
   }
 
@@ -1094,6 +1118,7 @@ class Device extends EventEmitter {
   private _onSignalingOffline = () => {
     this._log.info('Stream is offline');
     this._status = Device.Status.Offline;
+    this._edge = null;
     this._region = null;
     this.emit('offline', this);
   }
@@ -1472,6 +1497,16 @@ namespace Device {
     dscp?: boolean;
 
     /**
+     * The edge value corresponds to the geographic location that the client
+     * will use to connect to Twilio infrastructure. The default value is
+     * "roaming" which automatically selects an edge based on the latency of the
+     * client relative to available edges. You may not specify both `edge` and
+     * `region` in the Device options. Specifying both `edge` and `region` will
+     * result in an `InvalidArgumentException`.
+     */
+    edge?: string;
+
+    /**
      * Whether to automatically restart ICE when media connection fails
      */
     enableIceRestart?: boolean;
@@ -1507,6 +1542,35 @@ namespace Device {
 
     /**
      * The region code of the region to connect to.
+     *
+     * @deprecated
+     *
+     * CLIENT-7519 This parameter is deprecated in favor of the `edge`
+     * parameter. You may not specify both `edge` and `region` in the Device
+     * options.
+     *
+     * This parameter will be removed in the next major version release.
+     *
+     * The following table lists the new edge names to region name mappings.
+     * Instead of passing the `region` value in `options.region`, please pass the
+     * following `edge` value in `options.edge`.
+     *
+     * | Region Value | Edge Value   |
+     * |:-------------|:-------------|
+     * | au1          | sydney       |
+     * | br1          | sao-paolo    |
+     * | ie1          | dublin       |
+     * | de1          | frankfurt    |
+     * | jp1          | tokyo        |
+     * | sg1          | singapore    |
+     * | us1          | ashburn      |
+     * | us2          | umatilla     |
+     * | gll          | roaming      |
+     * | us1-ix       | ashburn-ix   |
+     * | us2-ix       | san-jose-ix  |
+     * | ie1-ix       | london-ix    |
+     * | de1-ix       | frankfurt-ix |
+     * | sg1-ix       | singapore-ix |
      */
     region?: string;