+++ tests/unit/device.ts
@@ -1,8 +1,12 @@
 import { levels as LogLevels } from 'loglevel';
 import Connection from '../../lib/twilio/connection';
 import Device from '../../lib/twilio/device';
-import { regionShortcodes } from '../../lib/twilio/regions';
 import { GeneralErrors } from '../../lib/twilio/errors';
+import {
+  Region,
+  regionShortcodes,
+  regionToEdge,
+} from '../../lib/twilio/regions';
 
 import * as assert from 'assert';
 import { EventEmitter } from 'events';
@@ -227,6 +231,11 @@ describe('Device', function() {
           assert.throws(() => (device.setup as any)(), /Token is required/);
         });
 
+        it('should throw if both `edge` and `region` are defined in options', () => {
+          device = new Device();
+          assert.throws(() => device.setup(token, { edge: 'foo', region: 'bar' }));
+        });
+
         it('should set device.isInitialized to true', () => {
           assert.equal(device.isInitialized, true);
         });
@@ -416,7 +425,40 @@ describe('Device', function() {
       });
     });
 
+    describe('.edge', () => {
+      it(`should return 'null' if not connected`, () => {
+        assert.equal(device.edge, null);
+      });
+
+      // these unit tests will need to be changed for Phase 2 Regional
+      context('when the region is mapped to a known edge', () => {
+        Object.entries(regionShortcodes).forEach(([fullName, region]: [string, string]) => {
+          const preferredEdge = regionToEdge[region as Region];
+          it(`should return ${preferredEdge} for ${region}`, () => {
+            pstream.emit('connected', { region: fullName });
+            assert.equal(device.edge, preferredEdge);
+          });
+        });
+      });
+
+      context('when the region is not mapped to a known edge', () => {
+        ['FOO_BAR', ''].forEach((name: string) => {
+          it(`should return the region string directly if it's '${name}'`, () => {
+            pstream.emit('connected', { region: name });
+            assert.equal(device.region(), name);
+          });
+        });
+      });
+    });
+
     describe('.region()', () => {
+      it(`should log.warn a deprecation warning`, () => {
+        const spy = sinon.spy();
+        device['_log'].warn = spy;
+        device.region();
+        assert(spy.calledOnce);
+      });
+
       it(`should return 'offline' if not connected`, () => {
         assert.equal(device.region(), 'offline');
       });